Add: - README.md - new flags (env/debug) Update: - debug mode for exec run command - default container run mode for normal call - fields on ContainerJob struct Remove: - comments - unused fields/values

+++ src/main.go
@@ -12,18 +12,28 @@ import (
 	"github.com/docker/docker/api/types/container"
 	"github.com/docker/docker/client"
 
-	// "github.com/docker/docker/pkg/stdcopy"
-
 	"os"
 )
 
 type ContainerJob struct {
-	ImageName       *string
-	SourcePath      *string
-	Commands        []string
+	ImageName       string
+	Command         string
+	Debug           bool
+	Env             multiValueFlag
 	ContainerObject container.ContainerCreateCreatedBody
 	Client          *client.Client
 	Error           error // keep track of latest error for container client
+}
+
+type multiValueFlag []string
+
+func (obj *multiValueFlag) String() string {
+	return strings.Join(*obj, " ")
+}
+
+func (obj *multiValueFlag) Set(s string) error {
+	*obj = append(*obj, s)
+	return nil
 }
 
 type actionrequest struct {
@@ -31,7 +41,7 @@ type actionrequest struct {
 }
 
 func (cjob ContainerJob) PullImage(ctx context.Context) {
-	reader, err := cjob.Client.ImagePull(ctx, *cjob.ImageName, types.ImagePullOptions{})
+	reader, err := cjob.Client.ImagePull(ctx, cjob.ImageName, types.ImagePullOptions{})
 	if err != nil {
 		panic(err)
 	}
@@ -42,10 +52,10 @@ func (cjob ContainerJob) PullImage(ctx context.Context) {
 
 func (cjob *ContainerJob) CreateJob(ctx context.Context) {
 	obj, err := cjob.Client.ContainerCreate(ctx, &container.Config{
-		Image: *cjob.ImageName,
+		Image: cjob.ImageName,
 		Tty:   false,
-		Cmd:   []string{"tail", "-f", "/dev/stdout"},
-		Env:   []string{"GOPATH", "GO111MODULE=auto"},
+		Cmd:   []string{"tail", "-f", "/dev/null"},
+		Env:   cjob.Env,
 	}, &container.HostConfig{
 		Binds: []string{
 			fmt.Sprintf("%s:/go", os.Getenv("PWD")),
@@ -59,7 +69,44 @@ func (cjob *ContainerJob) CreateJob(ctx context.Context) {
 	cjob.Error = cjob.Client.ContainerStart(ctx, cjob.ContainerObject.ID, types.ContainerStartOptions{})
 
 	cjob.ContainerError()
-	fmt.Printf("[%s] Job with image '%s' started successfully\n", cjob.ContainerObject.ID[0:12], *cjob.ImageName)
+	fmt.Printf("[%s] Job with image '%s' started successfully\n", cjob.ContainerObject.ID[0:12], cjob.ImageName)
+
+}
+
+func (cjob *ContainerJob) CreateJobWithInput(ctx context.Context) {
+
+	obj, err := cjob.Client.ContainerCreate(ctx, &container.Config{
+		AttachStdout: true,
+		AttachStderr: true,
+		Image:        cjob.ImageName,
+		Tty:          false,
+		Cmd:          []string{"/bin/sh", "-c", cjob.Command},
+		Env:          cjob.Env,
+	}, &container.HostConfig{
+		Binds: []string{
+			fmt.Sprintf("%s:/go", os.Getenv("PWD")),
+		},
+	}, nil, nil, "")
+
+	cjob.ContainerObject = obj
+	cjob.Error = err
+	cjob.ContainerError()
+
+	cjob.Error = cjob.Client.ContainerStart(ctx, cjob.ContainerObject.ID, types.ContainerStartOptions{})
+
+	cjob.ContainerError()
+
+	fmt.Printf("[%s] Job with image '%s' ran successfully\n", cjob.ContainerObject.ID[0:12], cjob.ImageName)
+
+	reader, err := cjob.Client.ContainerLogs(ctx, cjob.ContainerObject.ID, types.ContainerLogsOptions{ShowStdout: true, ShowStderr: true, Follow: true})
+
+	cjob.Error = err
+	cjob.ContainerError()
+
+	defer reader.Close()
+	io.Copy(os.Stdout, reader)
+
+	fmt.Printf("[%s] Job with image '%s' finished successfully\n", cjob.ContainerObject.ID[0:12], cjob.ImageName)
 
 }
 
@@ -80,33 +127,40 @@ func (cjob ContainerJob) DeleteJob(ctx context.Context) {
 
 }
 
-func (cjob ContainerJob) ExecJob(ctx context.Context, action string) {
-	fmt.Printf("[%s] Executing '%s' action\n", cjob.ContainerObject.ID[0:12], action)
+func (cjob ContainerJob) ExecJob(ctx context.Context, cmd string) {
+	fmt.Printf("[%s] Executing '%s' action\n", cjob.ContainerObject.ID[0:12], cmd)
 	exec, err := cjob.Client.ContainerExecCreate(ctx, cjob.ContainerObject.ID, types.ExecConfig{
-		Detach: true,
-		Cmd:    []string{"/bin/sh", "-c", action},
+		AttachStderr: true,
+		AttachStdout: true,
+		Tty:          true,
+		Cmd:          []string{"/bin/sh", "-c", cmd},
 	})
 
 	cjob.Error = err
 	cjob.ContainerError()
 
-	err = cjob.Client.ContainerExecStart(ctx, exec.ID, types.ExecStartCheck{
-		Detach: true,
-		Tty:    true,
-	})
+	attach, err := cjob.Client.ContainerExecAttach(ctx, exec.ID, types.ExecStartCheck{})
 
 	cjob.Error = err
 	cjob.ContainerError()
 
-	fmt.Printf("[%s] Action '%s' executed succesfully\n", cjob.ContainerObject.ID[0:12], action)
+	defer attach.Close()
+	io.Copy(os.Stdout, attach.Reader)
+
+	fmt.Printf("[%s] Action '%s' executed succesfully\n", cjob.ContainerObject.ID[0:12], cmd)
 }
 
 func (cjob *ContainerJob) SetFields(a actionrequest) {
-	cjob.ImageName = a.Flag.String("image", "golang:1.18.0-alpine3.15", "name of image used for build command")
-	cjob.SourcePath = a.Flag.String("source-path", "", "path of the directory or file you want to run the jobs on")
+	a.Flag.StringVar(&cjob.ImageName, "image", "golang:1.18.0-alpine3.15", "name of image used for build command")
+	a.Flag.BoolVar(&cjob.Debug, "d", false, "create and exec is used over run")
+	a.Flag.Var(&cjob.Env, "env", "environment variables to be passed")
+	a.Flag.Var(&cjob.Env, "e", "environment variables to be passed")
+
 	a.Flag.Parse(os.Args[3:])
-	cjob.Commands = strings.Split(strings.Join(a.Flag.Args(), " "), ",")
-	fmt.Println("Commands: %v", cjob.Commands)
+	cjob.Command = strings.Join(a.Flag.Args(), " ")
+
+	fmt.Printf("Command: %s\n", cjob.Command)
+	fmt.Println(cjob.Env.String())
 }
 
 func (a actionrequest) BuildSubCommandExecute() {
@@ -121,23 +175,20 @@ func (a actionrequest) BuildSubCommandExecute() {
 	cjob.ContainerError()
 
 	cjob.PullImage(ctx)
-	cjob.CreateJob(ctx)
 
-	// for loop with execs
-	// cjob.ExecJob(ctx, cjob.Commands)
-	for _, exec := range cjob.Commands {
-		cjob.ExecJob(ctx, exec)
+	if cjob.Debug {
+		cjob.CreateJob(ctx)
+
+		// for loop with execs
+		for _, exec := range strings.Split(cjob.Command, "&&") {
+			cjob.ExecJob(ctx, strings.TrimRight(strings.TrimLeft(exec, " "), " "))
+		}
+
+		// stop container
+		cjob.DeleteJob(ctx)
+	} else {
+		cjob.CreateJobWithInput(ctx)
 	}
-
-	reader, err := cjob.Client.ContainerLogs(ctx, cjob.ContainerObject.ID, types.ContainerLogsOptions{ShowStdout: true})
-
-	cjob.Error = err
-	cjob.ContainerError()
-
-	defer reader.Close()
-	io.Copy(os.Stdout, reader)
-
-	cjob.DeleteJob(ctx)
 }
 
 func (a actionrequest) Execute() error {
